step 3: activate (#60)
* step 3: activate

This uncomments the `activation` section of the values.yaml, resulting
of creation of an activation job which activtes the chain. It also sets
the number of faucets back to 10000 so the faucet server will start.

This has to go `up` after the activation date and time in the
values.yaml file. Otherwise, activation will fail.

If activation fails for any reason, it's always possible to delete the
job, then `up --refresh` will create a fresh job that will attempt
activation again.

But activation is irreversible - once a block has been out and
propagated to peers, it's too late. Measure twice, cut once.

* Add julien non bootstrap baker account

* Add bakingbadfaucet account

Co-authored-by: Aryeh Harris <harryttd@gmail.com>